authentication_completed

+++ controller/expense.js
@@ -7,7 +7,7 @@ const addExpense = (req, res) => {
     //     return res.status(400).json({success: false, message: 'Parameters missing'})
     // }
     
-    Expense.create({ expenseamount, description, category}).then(expense => {
+    Expense.create({ expenseamount, description, category,userId: req.user.id}).then(expense => {
         return res.status(201).json({expense, success: true } );
     }).catch(err => {
         return res.status(500).json({success : false, error: err})
@@ -15,6 +15,7 @@ const addExpense = (req, res) => {
 }
 
 const getAllExpenses = (req, res) => {
+    //console.log(res.body);
     Expense.findAll({where : {userId: req.user.id}})
         .then(expenses => {
             return res.status(200).json({ expenses, success: true });
@@ -29,7 +30,7 @@ const getAllExpenses = (req, res) => {
 const deleteExpense = (req, res) => {
     const expenseId = req.params.id;
   
-    Expense.destroy({ where: { id: expenseId } })
+    Expense.destroy({ where: { id: expenseId ,userId: req.user.id} })
       .then(() => {
         res.status(204).end(); // Respond with a 204 status code (No Content) to indicate success
       })
